requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
alembic==1.17.2
    # via optuna
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.6
    # via fastmcp
beartype==0.22.9
    # via
    #   py-key-value-aio
    #   py-key-value-shared
cachetools==6.2.4
    # via py-key-value-aio
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
cloudpickle==3.1.2
    # via pydocket
colorlog==6.10.1
    # via optuna
coverage==7.11.3
    # via pytest-cov
cryptography==46.0.3
    # via
    #   authlib
    #   pyjwt
cyclopts==4.4.3
    # via fastmcp
cython==3.2.0
    # via -r requirements.in
diskcache==5.6.3
    # via py-key-value-aio
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.4
    # via rich-rst
email-validator==2.3.0
    # via pydantic
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.3.1
    # via fastmcp
fakeredis==2.33.0
    # via pydocket
fastapi==0.128.0
    # via -r requirements.in
fastmcp==2.14.2
    # via -r requirements.in
greenlet==3.3.0
    # via
    #   playwright
    #   sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
iniconfig==2.3.0
    # via pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.2
    # via keyring
jaraco-functools==4.4.0
    # via keyring
joblib==1.5.3
    # via scikit-learn
jsonschema==4.25.1
    # via mcp
jsonschema-path==0.3.4
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.7.0
    # via py-key-value-aio
linkify-it-py==2.0.3
    # via markdown-it-py
lupa==2.6
    # via fakeredis
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
markupsafe==3.0.3
    # via mako
mcp==1.25.0
    # via
    #   -r requirements.in
    #   fastmcp
mdit-py-plugins==0.5.0
    # via textual
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via
    #   jaraco-classes
    #   jaraco-functools
nats-py==2.12.0
    # via -r requirements.in
networkx==3.6.1
    # via -r requirements.in
numpy==2.3.4
    # via
    #   -r requirements.in
    #   optuna
    #   pandas
    #   scikit-learn
    #   scipy
    #   xgboost
openapi-pydantic==0.5.1
    # via fastmcp
openpyxl==3.1.5
    # via -r requirements.in
opentelemetry-api==1.39.1
    # via
    #   opentelemetry-exporter-prometheus
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydocket
opentelemetry-exporter-prometheus==0.60b1
    # via pydocket
opentelemetry-instrumentation==0.60b1
    # via pydocket
opentelemetry-sdk==1.39.1
    # via opentelemetry-exporter-prometheus
opentelemetry-semantic-conventions==0.60b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
optuna==4.6.0
    # via -r requirements.in
packaging==25.0
    # via
    #   opentelemetry-instrumentation
    #   optuna
    #   pytest
pandas==2.3.3
    # via -r requirements.in
pathable==0.4.4
    # via jsonschema-path
pathvalidate==3.3.1
    # via py-key-value-aio
platformdirs==4.5.1
    # via
    #   fastmcp
    #   textual
playwright==1.57.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
prometheus-client==0.23.1
    # via
    #   opentelemetry-exporter-prometheus
    #   pydocket
py-key-value-aio==0.3.0
    # via
    #   fastmcp
    #   pydocket
py-key-value-shared==0.3.0
    # via py-key-value-aio
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   fastmcp
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pydocket==0.16.3
    # via fastmcp
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via
    #   pytest
    #   rich
    #   textual
pyjwt==2.10.1
    # via mcp
pyperclip==1.11.0
    # via fastmcp
pytest==9.0.0
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==7.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   fastmcp
    #   pydantic-settings
    #   uvicorn
python-json-logger==4.0.0
    # via pydocket
python-multipart==0.0.21
    # via mcp
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   jsonschema-path
    #   optuna
    #   uvicorn
redis==7.1.0
    # via
    #   fakeredis
    #   py-key-value-aio
    #   pydocket
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   jsonschema-path
rich==14.2.0
    # via
    #   cyclopts
    #   fastmcp
    #   pydocket
    #   rich-rst
    #   textual
    #   typer
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.15.4
    # via -r requirements.in
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.16.3
    # via
    #   scikit-learn
    #   xgboost
setuptools==80.9.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via fakeredis
sqlalchemy==2.0.45
    # via
    #   alembic
    #   optuna
sse-starlette==3.1.2
    # via mcp
starlette==0.50.0
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
textual==7.0.0
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via optuna
typer==0.21.0
    # via pydocket
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   mcp
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   py-key-value-shared
    #   pydantic
    #   pydantic-core
    #   pydocket
    #   pyee
    #   sqlalchemy
    #   textual
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
uvicorn==0.40.0
    # via
    #   -r requirements.in
    #   fastmcp
    #   mcp
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   fastmcp
    #   uvicorn
wrapt==1.17.3
    # via opentelemetry-instrumentation
xgboost==3.1.2
    # via -r requirements.in
xlrd==2.0.2
    # via -r requirements.in
zipp==3.23.0
    # via importlib-metadata
